Add files via upload
A To-Do List application is a useful project that helps users
manage and organize their tasks efficiently. This project aims to
create a command-line or GUI-based application using Python,
allowing users to create, update, and track their to-do lists


Tkinter module : 

Tkinter is the inbuilt python module that is used to create GUI applications. It is one of the most commonly used modules for creating GUI applications in Python as it is simple and easy to work with.

Basic Tkinter components: 

pack() : It is used to centralized the buttons on the output window page.we can even change position of the buttons at left or right.

configure() : It is used to set the background color of the screen page.

messagebox : It is used to generate a popup message which can contain the warnning/information/error if any unwanted condition occurs.

Entry() : It is used to form a text box in which the user can write.